🚀 feat: add support for configurable table name in database
The update includes changes to support a configurable table name in the database. The table name is now retrieved from the `api_keys` module and used in various SQL queries throughout the code. This allows for greater flexibility and easier customization of the database structure.
Also added favorite to function to eliminate 3 types copy.